fix(assetlibrary-export): refactored batcherBase
